Split JSON and YAML serializer templates
